This change fixes the warning generated by clang when we enable -Wobjc-property-assign-on-object-type flag under the ARC environment.
Clang with the flag generates the warning if a property to an Objective-C object/block has an assign attribute because an assign property is treated as unretained property and it potentially becomes a dangling pointer.
The flag is added by https://github.com/llvm/llvm-project/commit/52a503d4f333d
So using strong / weak / copy would be better in general.

selectedViewController is the currently selected view controller and the view must be in viewControllers which is strongly retained.
I think it's safe to use weak to selectedViewController.

PiperOrigin-RevId: 755861552